+++ src/app/profile/my-profile/page.tsx
@@ -1,10 +1,11 @@
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/common/configs/nextAuth.config';
+
 import BaseButton from '@/common/components/Buttons/BaseButton';
 import Divider from '@/common/components/Divider';
 import PageTitle from '@/common/components/PageTitle';
 import ProfileContainer from '@/app/profile/my-profile/_components/ProfileContainer';
 import ProfileSubMenu from '@/app/profile/my-profile/_components/ProfileSubMenu';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
-import { getServerSession } from 'next-auth';
 import PasswordInput from '@/common/components/Inputs/PasswordInput';
 
 export default async function myProfilePage() {
