+++ src/common/layouts/Header/Header.tsx
@@ -1,9 +1,10 @@
+import Link from 'next/link';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/common/configs/nextAuth.config';
+
 import { Bars3Icon } from '@heroicons/react/24/solid';
 import SearchWrapper from '@/common/components/Search/SearchWrapper';
-import Link from 'next/link';
 import UserState from '@/common/components/UserState';
-import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
 
 export default async function Header() {
 	const session = await getServerSession(authOptions);
